pyproject.toml
[tool.poetry]
name = "dev-dotenv"
version = "0.0.1"
description = ""
authors = ["Janko Thyson <janko@kaosmaps.com>"]
readme = "README.md"
packages = [{include = "dev_dotenv", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv-vault = "^0.6.4"
streamlit = "^1.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev-dotenv = "dev_dotenv.cli:cli"
